refactor(sandbox-docker): split files to fit complexity budget
Extract persistence flow into find-or-create.ts (doFindOrCreate,
doDestroyScope, tryReuse helpers) and split default-client.ts into
default-client-helpers.ts (subprocess + parsing utilities) +
default-client-ops.ts (per-method handlers). Pulls factory bodies
out of scope-registry.ts so the module-level functions own state.

Drops 6 complexity violations introduced by the persistence work
(adapter.ts file size, default-client.ts file size, tryReuse,
createDefaultDockerClient, makeContainer, createFileScopeRegistry).

No behavior change — all 165 existing tests still pass.
